Create Iphone sales analysis
Apple iPhones are among the top-selling smartphones worldwide. There is huge competition among smartphone brands in India, where you can get the latest technology in a smartphone at half the price of an iPhone. Still, there are high sales of iPhones in India.Lets Analyse iphone sales in India with the flipkart data.¶